bot_scripts/requirements.txt
python-telegram-bot==13.7
PyGithub==1.55
matplotlib==3.5.1